管易商品类接口对接

+++ erpprovider-gy/src/main/java/com/huobanplus/erpprovider/gy/formatgy/order/GYResponseOrderItem.java
@@ -19,7 +19,7 @@ public class GYResponseOrderItem {
      * 退款状态
      */
     @JSONField(name= "refund")
-    private double refund;
+    private int refund;
 
     /**
      * 商品代码
